requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu
psycopg2-binary # PostgreSQL bağlantısı için
SQLAlchemy